AdS5_3_1: move BulkHorizon and HorizonCache to types
and add a parameter for the FD order of the xi_t equation
