+++ src/AdS5_3_1/AdS5_3_1.jl
@@ -1,6 +1,8 @@
 module AdS5_3_1
 
 using Jecco
+using Jecco: mul_col!
+using SparseArrays: SparseMatrixCSC
 using OrdinaryDiffEq
 
 # abstract types and structs used throughout
