+++ src/AdS5_3_1/compute_gauge_t.jl
@@ -1,154 +1,4 @@
 
-using Jecco: mul_col!
-using SparseArrays: SparseMatrixCSC
-
-struct BulkHorizon{T}
-    B1_uAH      :: Array{T,3}
-    B2_uAH      :: Array{T,3}
-    G_uAH       :: Array{T,3}
-    phi_uAH     :: Array{T,3}
-    S_uAH       :: Array{T,3}
-    Fx_uAH      :: Array{T,3}
-    Fy_uAH      :: Array{T,3}
-    Sd_uAH      :: Array{T,3}
-    B1d_uAH     :: Array{T,3}
-    B2d_uAH     :: Array{T,3}
-    Gd_uAH      :: Array{T,3}
-    phid_uAH    :: Array{T,3}
-    A_uAH       :: Array{T,3}
-
-    Du_B1_uAH   :: Array{T,3}
-    Du_B2_uAH   :: Array{T,3}
-    Du_G_uAH    :: Array{T,3}
-    Du_phi_uAH  :: Array{T,3}
-    Du_S_uAH    :: Array{T,3}
-    Du_Fx_uAH   :: Array{T,3}
-    Du_Fy_uAH   :: Array{T,3}
-    Du_Sd_uAH   :: Array{T,3}
-    Du_B1d_uAH  :: Array{T,3}
-    Du_B2d_uAH  :: Array{T,3}
-    Du_Gd_uAH   :: Array{T,3}
-    Du_A_uAH    :: Array{T,3}
-
-    Duu_B1_uAH  :: Array{T,3}
-    Duu_B2_uAH  :: Array{T,3}
-    Duu_G_uAH   :: Array{T,3}
-    Duu_S_uAH   :: Array{T,3}
-    Duu_Fx_uAH  :: Array{T,3}
-    Duu_Fy_uAH  :: Array{T,3}
-    Duu_A_uAH   :: Array{T,3}
-end
-function BulkHorizon{T}(Nx::Int, Ny::Int) where {T<:Real}
-    B1_uAH      = Array{T}(undef, 1, Nx, Ny)
-    B2_uAH      = Array{T}(undef, 1, Nx, Ny)
-    G_uAH       = Array{T}(undef, 1, Nx, Ny)
-    phi_uAH     = Array{T}(undef, 1, Nx, Ny)
-    S_uAH       = Array{T}(undef, 1, Nx, Ny)
-    Fx_uAH      = Array{T}(undef, 1, Nx, Ny)
-    Fy_uAH      = Array{T}(undef, 1, Nx, Ny)
-    Sd_uAH      = Array{T}(undef, 1, Nx, Ny)
-    B1d_uAH     = Array{T}(undef, 1, Nx, Ny)
-    B2d_uAH     = Array{T}(undef, 1, Nx, Ny)
-    Gd_uAH      = Array{T}(undef, 1, Nx, Ny)
-    phid_uAH    = Array{T}(undef, 1, Nx, Ny)
-    A_uAH       = Array{T}(undef, 1, Nx, Ny)
-
-    Du_B1_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_B2_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_G_uAH    = Array{T}(undef, 1, Nx, Ny)
-    Du_phi_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Du_S_uAH    = Array{T}(undef, 1, Nx, Ny)
-    Du_Fx_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_Fy_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_Sd_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_B1d_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Du_B2d_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Du_Gd_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Du_A_uAH    = Array{T}(undef, 1, Nx, Ny)
-
-    Duu_B1_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Duu_B2_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Duu_G_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Duu_S_uAH   = Array{T}(undef, 1, Nx, Ny)
-    Duu_Fx_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Duu_Fy_uAH  = Array{T}(undef, 1, Nx, Ny)
-    Duu_A_uAH   = Array{T}(undef, 1, Nx, Ny)
-
-    BulkHorizon{T}(B1_uAH, B2_uAH, G_uAH, phi_uAH, S_uAH, Fx_uAH, Fy_uAH,
-                   Sd_uAH, B1d_uAH, B2d_uAH, Gd_uAH, phid_uAH, A_uAH, Du_B1_uAH,
-                   Du_B2_uAH, Du_G_uAH, Du_phi_uAH, Du_S_uAH, Du_Fx_uAH,
-                   Du_Fy_uAH, Du_Sd_uAH, Du_B1d_uAH, Du_B2d_uAH, Du_Gd_uAH,
-                   Du_A_uAH, Duu_B1_uAH, Duu_B2_uAH, Duu_G_uAH, Duu_S_uAH,
-                   Duu_Fx_uAH, Duu_Fy_uAH, Duu_A_uAH)
-end
-
-struct HorizonCache{T,D}
-    bulkhorizon :: BulkHorizon{T}
-    axx         :: Vector{T}
-    ayy         :: Vector{T}
-    axy         :: Vector{T}
-    bx          :: Vector{T}
-    by          :: Vector{T}
-    cc          :: Vector{T}
-    b_vec       :: Vector{T}
-    Dx_2D       :: D
-    Dy_2D       :: D
-    Dxx_2D      :: D
-    Dyy_2D      :: D
-    Dxy_2D      :: D
-    _Dx_2D      :: D
-    _Dy_2D      :: D
-    _Dxx_2D     :: D
-    _Dyy_2D     :: D
-    _Dxy_2D     :: D
-end
-function HorizonCache(sys::System, ord::Int)
-    _, Nx, Ny = size(sys)
-    hx = Jecco.delta(sys.xcoord)
-    hy = Jecco.delta(sys.ycoord)
-    T  = Jecco.coord_eltype(sys.ucoord)
-    M  = Nx * Ny
-
-    bulkhorizon = BulkHorizon{T}(Nx, Ny)
-
-    axx    = Vector{T}(undef, M)
-    ayy    = Vector{T}(undef, M)
-    axy    = Vector{T}(undef, M)
-    bx     = Vector{T}(undef, M)
-    by     = Vector{T}(undef, M)
-    cc     = Vector{T}(undef, M)
-    b_vec  = Vector{T}(undef, M)
-
-    Dx_    = CenteredDiff{1}(1, ord, hx, Nx)
-    Dxx_   = CenteredDiff{1}(2, ord, hx, Nx)
-    Dy_    = CenteredDiff{2}(1, ord, hy, Ny)
-    Dyy_   = CenteredDiff{2}(2, ord, hy, Ny)
-
-    #=
-    use the Kronecker product (kron) to build 2-dimensional derivation matrices
-    from 1-dimensional ones. see for instance:
-
-    https://en.wikipedia.org/wiki/Kronecker_product
-    https://arxiv.org/pdf/1801.01483.pdf (section 5)
-    =#
-    Dx_2D  = kron(I(Ny), SparseMatrixCSC(Dx_))
-    Dxx_2D = kron(I(Ny), SparseMatrixCSC(Dxx_))
-    Dy_2D  = kron(SparseMatrixCSC(Dy_), I(Nx))
-    Dyy_2D = kron(SparseMatrixCSC(Dyy_), I(Nx))
-    Dxy_2D = Dx_2D * Dy_2D
-
-    _Dx_2D  = copy(Dx_2D)
-    _Dxx_2D = copy(Dxx_2D)
-    _Dy_2D  = copy(Dy_2D)
-    _Dyy_2D = copy(Dyy_2D)
-    _Dxy_2D = copy(Dxy_2D)
-
-    HorizonCache{T,typeof(Dx_2D)}(bulkhorizon, axx, ayy, axy, bx, by, cc, b_vec,
-                                  Dx_2D,  Dxx_2D,  Dy_2D,  Dyy_2D,  Dxy_2D,
-                                  _Dx_2D, _Dxx_2D, _Dy_2D, _Dyy_2D, _Dxy_2D)
-end
-
-
 function compute_xi_t!(gauge_t::Gauge, bulkconstrain::BulkConstrained,
                        bulkevol::BulkEvolved, deriv::BulkDeriv, gauge::Gauge,
                        cache::HorizonCache, sys::System{Outer}, ::ConstantGauge)
