+++ src/AdS5_3_1/run.jl
@@ -87,6 +87,7 @@ function run_model(grid::SpecCartGrid3D, id::InitialData, evoleq::EvolutionEquat
     bulkevols      = BulkEvolvedPartition(grid)
     bulkconstrains = BulkConstrainedPartition(grid)
     bulkderivs     = BulkDerivPartition(grid)
+    horizoncache   = HorizonCache(systems[end], evoleq.gaugecondition.fd_order)
 
     # initial conditions
     init_data!(bulkevols, boundary, gauge, systems, id)
@@ -95,7 +96,7 @@ function run_model(grid::SpecCartGrid3D, id::InitialData, evoleq::EvolutionEquat
     evolvars  = EvolVars(boundary, gauge, bulkevols)
 
     # function that updates the state vector
-    rhs! = setup_rhs(bulkconstrains, bulkderivs, systems)
+    rhs! = setup_rhs(bulkconstrains, bulkderivs, horizoncache, systems)
 
     dt0  = integration.dt
     tmax = integration.tmax
