+++ src/AdS5_3_1/system.jl
@@ -209,3 +209,49 @@ function BulkDerivPartition(grid::SpecCartGrid3D{T}) where {T}
 
     BulkPartition((bulk_in, bulk_out...))
 end
+
+function HorizonCache(sys::System, ord::Int)
+    _, Nx, Ny = size(sys)
+    hx = Jecco.delta(sys.xcoord)
+    hy = Jecco.delta(sys.ycoord)
+    T  = Jecco.coord_eltype(sys.ucoord)
+    M  = Nx * Ny
+
+    bulkhorizon = BulkHorizon{T}(Nx, Ny)
+
+    axx    = Vector{T}(undef, M)
+    ayy    = Vector{T}(undef, M)
+    axy    = Vector{T}(undef, M)
+    bx     = Vector{T}(undef, M)
+    by     = Vector{T}(undef, M)
+    cc     = Vector{T}(undef, M)
+    b_vec  = Vector{T}(undef, M)
+
+    Dx_    = CenteredDiff{1}(1, ord, hx, Nx)
+    Dxx_   = CenteredDiff{1}(2, ord, hx, Nx)
+    Dy_    = CenteredDiff{2}(1, ord, hy, Ny)
+    Dyy_   = CenteredDiff{2}(2, ord, hy, Ny)
+
+    #=
+    use the Kronecker product (kron) to build 2-dimensional derivation matrices
+    from 1-dimensional ones. see for instance:
+
+    https://en.wikipedia.org/wiki/Kronecker_product
+    https://arxiv.org/pdf/1801.01483.pdf (section 5)
+    =#
+    Dx_2D  = kron(I(Ny), SparseMatrixCSC(Dx_))
+    Dxx_2D = kron(I(Ny), SparseMatrixCSC(Dxx_))
+    Dy_2D  = kron(SparseMatrixCSC(Dy_), I(Nx))
+    Dyy_2D = kron(SparseMatrixCSC(Dyy_), I(Nx))
+    Dxy_2D = Dx_2D * Dy_2D
+
+    _Dx_2D  = copy(Dx_2D)
+    _Dxx_2D = copy(Dxx_2D)
+    _Dy_2D  = copy(Dy_2D)
+    _Dyy_2D = copy(Dyy_2D)
+    _Dxy_2D = copy(Dxy_2D)
+
+    HorizonCache{T,typeof(Dx_2D)}(bulkhorizon, axx, ayy, axy, bx, by, cc, b_vec,
+                                  Dx_2D,  Dxx_2D,  Dy_2D,  Dyy_2D,  Dxy_2D,
+                                  _Dx_2D, _Dxx_2D, _Dy_2D, _Dyy_2D, _Dxy_2D)
+end
